% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{associate_fleet}
\alias{associate_fleet}
\title{Associates the specified fleet with the specified stack}
\usage{
associate_fleet(FleetName, StackName)
}
\arguments{
\item{FleetName}{The name of the fleet.}

\item{StackName}{The name of the stack.}
}
\description{
Associates the specified fleet with the specified stack.
}
\examples{

}
