% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{batch_disassociate_user_stack}
\alias{batch_disassociate_user_stack}
\title{Disassociates the specified users from the specified stacks}
\usage{
batch_disassociate_user_stack(UserStackAssociations)
}
\arguments{
\item{UserStackAssociations}{[required] The list of UserStackAssociation objects.}
}
\description{
Disassociates the specified users from the specified stacks.
}
\section{Accepted Parameters}{
\preformatted{batch_disassociate_user_stack(
  UserStackAssociations = list(
    list(
      StackName = "string",
      UserName = "string",
      AuthenticationType = "API"|"SAML"|"USERPOOL",
      SendEmailNotification = TRUE|FALSE
    )
  )
)
}
}

