% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{create_directory_config}
\alias{create_directory_config}
\title{Creates a Directory Config object in AppStream 2}
\usage{
create_directory_config(DirectoryName, OrganizationalUnitDistinguishedNames,
  ServiceAccountCredentials)
}
\arguments{
\item{DirectoryName}{The fully qualified name of the directory (for example, corp.example.com).}

\item{OrganizationalUnitDistinguishedNames}{The distinguished names of the organizational units for computer accounts.}

\item{ServiceAccountCredentials}{The credentials for the service account used by the streaming instance to connect to the directory.}
}
\description{
Creates a Directory Config object in AppStream 2.0. This object includes the information required to join streaming instances to an Active Directory domain.
}
\examples{

}
