% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{create_image_builder_streaming_url}
\alias{create_image_builder_streaming_url}
\title{Creates a URL to start an image builder streaming session}
\usage{
create_image_builder_streaming_url(
  Name = "string",
  Validity = 123
)
}
\arguments{
\item{Name}{[required] The name of the image builder.}

\item{Validity}{The time that the streaming URL will be valid, in seconds. Specify a value between 1 and 604800 seconds. The default is 3600 seconds.}
}
\description{
Creates a URL to start an image builder streaming session.
}
\examples{

}
