% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{create_stack}
\alias{create_stack}
\title{Creates a stack to start streaming applications to users}
\usage{
create_stack(
  Name = "string",
  Description = "string",
  DisplayName = "string",
  StorageConnectors = list(
    list(
      ConnectorType = "HOMEFOLDERS"|"GOOGLE_DRIVE"|"ONE_DRIVE",
      ResourceIdentifier = "string",
      Domains = list(
        "string"
      )
    )
  ),
  RedirectURL = "string",
  FeedbackURL = "string",
  UserSettings = list(
    list(
      Action = "CLIPBOARD_COPY_FROM_LOCAL_DEVICE"|"CLIPBOARD_COPY_TO_LOCAL_DEVICE"|"FILE_UPLOAD"|"FILE_DOWNLOAD"|"PRINTING_TO_LOCAL_DEVICE",
      Permission = "ENABLED"|"DISABLED"
    )
  ),
  ApplicationSettings = list(
    Enabled = TRUE|FALSE,
    SettingsGroup = "string"
  )
)
}
\arguments{
\item{Name}{[required] The name of the stack.}

\item{Description}{The description for display.}

\item{DisplayName}{The stack name for display.}

\item{StorageConnectors}{The storage connectors to enable.}

\item{RedirectURL}{The URL that users are redirected to after their streaming session ends.}

\item{FeedbackURL}{The URL that users are redirected to after they click the Send Feedback link. If no URL is specified, no Send Feedback link is displayed.}

\item{UserSettings}{The actions that are enabled or disabled for users during their streaming sessions. By default, these actions are enabled.}

\item{ApplicationSettings}{The persistent application settings for users of a stack. When these settings are enabled, changes that users make to applications and Windows settings are automatically saved after each session and applied to the next session.}
}
\description{
Creates a stack to start streaming applications to users. A stack consists of an associated fleet, user access policies, and storage configurations.
}
\examples{

}
