% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{delete_fleet}
\alias{delete_fleet}
\title{Deletes the specified fleet}
\usage{
delete_fleet(
  Name = "string"
)
}
\arguments{
\item{Name}{[required] The name of the fleet.}
}
\description{
Deletes the specified fleet.
}
\examples{

}
