% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{delete_image}
\alias{delete_image}
\title{Deletes the specified image}
\usage{
delete_image(
  Name = "string"
)
}
\arguments{
\item{Name}{[required] The name of the image.}
}
\description{
Deletes the specified image. You cannot delete an image when it is in use. After you delete an image, you cannot provision new capacity using the image.
}
\examples{

}
