% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{delete_image_builder}
\alias{delete_image_builder}
\title{Deletes the specified image builder and releases the capacity}
\usage{
delete_image_builder(
  Name = "string"
)
}
\arguments{
\item{Name}{[required] The name of the image builder.}
}
\description{
Deletes the specified image builder and releases the capacity.
}
\examples{

}
