% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{delete_image_permissions}
\alias{delete_image_permissions}
\title{Deletes permissions for the specified private image}
\usage{
delete_image_permissions(
  Name = "string",
  SharedAccountId = "string"
)
}
\arguments{
\item{Name}{[required] The name of the private image.}

\item{SharedAccountId}{[required] The 12-digit ID of the AWS account for which to delete image permissions.}
}
\description{
Deletes permissions for the specified private image. After you delete permissions for an image, AWS accounts to which you previously granted these permissions can no longer use the image.
}
\examples{

}
