% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{delete_stack}
\alias{delete_stack}
\title{Deletes the specified stack}
\usage{
delete_stack(Name)
}
\arguments{
\item{Name}{[required] The name of the stack.}
}
\description{
Deletes the specified stack. After the stack is deleted, the application streaming environment provided by the stack is no longer available to users. Also, any reservations made for application streaming sessions for the stack are released.
}
\section{Accepted Parameters}{
\preformatted{delete_stack(
  Name = "string"
)
}
}

