% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{delete_user}
\alias{delete_user}
\title{Deletes a user from the user pool}
\usage{
delete_user(
  UserName = "string",
  AuthenticationType = "API"|"SAML"|"USERPOOL"
)
}
\arguments{
\item{UserName}{[required] The email address of the user.}

\item{AuthenticationType}{[required] The authentication type for the user. You must specify USERPOOL.}
}
\description{
Deletes a user from the user pool.
}
\examples{

}
