% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{describe_directory_configs}
\alias{describe_directory_configs}
\title{Retrieves a list that describes one or more specified Directory Config objects for AppStream 2}
\usage{
describe_directory_configs(DirectoryNames = NULL, MaxResults = NULL,
  NextToken = NULL)
}
\arguments{
\item{DirectoryNames}{The directory names.}

\item{MaxResults}{The maximum size of each page of results.}

\item{NextToken}{The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.}
}
\description{
Retrieves a list that describes one or more specified Directory Config objects for AppStream 2.0, if the names for these objects are provided. Otherwise, all Directory Config objects in the account are described. These objects include the information required to join streaming instances to an Active Directory domain.
}
\details{
Although the response syntax in this topic includes the account password, this password is not returned in the actual response.
}
\section{Accepted Parameters}{
\preformatted{describe_directory_configs(
  DirectoryNames = list(
    "string"
  ),
  MaxResults = 123,
  NextToken = "string"
)
}
}

