% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{describe_fleets}
\alias{describe_fleets}
\title{Retrieves a list that describes one or more specified fleets, if the fleet names are provided}
\usage{
describe_fleets(
  Names = list(
    "string"
  ),
  NextToken = "string"
)
}
\arguments{
\item{Names}{The names of the fleets to describe.}

\item{NextToken}{The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.}
}
\description{
Retrieves a list that describes one or more specified fleets, if the fleet names are provided. Otherwise, all fleets in the account are described.
}
\examples{

}
