% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{describe_image_builders}
\alias{describe_image_builders}
\title{Retrieves a list that describes one or more specified image builders, if the image builder names are provided}
\usage{
describe_image_builders(
  Names = list(
    "string"
  ),
  MaxResults = 123,
  NextToken = "string"
)
}
\arguments{
\item{Names}{The names of the image builders to describe.}

\item{MaxResults}{The maximum size of each page of results.}

\item{NextToken}{The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.}
}
\description{
Retrieves a list that describes one or more specified image builders, if the image builder names are provided. Otherwise, all image builders in the account are described.
}
\examples{

}
