% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{describe_stacks}
\alias{describe_stacks}
\title{Retrieves a list that describes one or more specified stacks, if the stack names are provided}
\usage{
describe_stacks(Names = NULL, NextToken = NULL)
}
\arguments{
\item{Names}{The names of the stacks to describe.}

\item{NextToken}{The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.}
}
\description{
Retrieves a list that describes one or more specified stacks, if the stack names are provided. Otherwise, all stacks in the account are described.
}
\section{Accepted Parameters}{
\preformatted{describe_stacks(
  Names = list(
    "string"
  ),
  NextToken = "string"
)
}
}

