% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{disable_user}
\alias{disable_user}
\title{Disables the specified user in the user pool}
\usage{
disable_user(
  UserName = "string",
  AuthenticationType = "API"|"SAML"|"USERPOOL"
)
}
\arguments{
\item{UserName}{[required] The email address of the user.}

\item{AuthenticationType}{[required] The authentication type for the user. You must specify USERPOOL.}
}
\description{
Disables the specified user in the user pool. Users can't sign in to AppStream 2.0 until they are re-enabled. This action does not delete the user.
}
\examples{

}
