% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{enable_user}
\alias{enable_user}
\title{Enables a user in the user pool}
\usage{
enable_user(
  UserName = "string",
  AuthenticationType = "API"|"SAML"|"USERPOOL"
)
}
\arguments{
\item{UserName}{[required] The email address of the user.}

\item{AuthenticationType}{[required] The authentication type for the user. You must specify USERPOOL.}
}
\description{
Enables a user in the user pool. After being enabled, users can sign in to AppStream 2.0 and open applications from the stacks to which they are assigned.
}
\examples{

}
