% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{expire_session}
\alias{expire_session}
\title{Immediately stops the specified streaming session}
\usage{
expire_session(SessionId)
}
\arguments{
\item{SessionId}{[required] The ID of the streaming session.}
}
\description{
Immediately stops the specified streaming session.
}
\section{Accepted Parameters}{
\preformatted{expire_session(
  SessionId = "string"
)
}
}

