% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{list_associated_fleets}
\alias{list_associated_fleets}
\title{Retrieves the name of the fleet that is associated with the specified stack}
\usage{
list_associated_fleets(StackName, NextToken = NULL)
}
\arguments{
\item{StackName}{[required] The name of the stack.}

\item{NextToken}{The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.}
}
\description{
Retrieves the name of the fleet that is associated with the specified stack.
}
\section{Accepted Parameters}{
\preformatted{list_associated_fleets(
  StackName = "string",
  NextToken = "string"
)
}
}

