% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{list_associated_stacks}
\alias{list_associated_stacks}
\title{Retrieves the name of the stack with which the specified fleet is associated}
\usage{
list_associated_stacks(
  FleetName = "string",
  NextToken = "string"
)
}
\arguments{
\item{FleetName}{[required] The name of the fleet.}

\item{NextToken}{The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.}
}
\description{
Retrieves the name of the stack with which the specified fleet is associated.
}
\examples{

}
