% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_package.R
\docType{package}
\name{paws.appstream-package}
\alias{paws.appstream}
\alias{paws.appstream-package}
\title{paws.appstream: Amazon AppStream}
\description{
You can use Amazon AppStream 2.0 to stream desktop
applications to any device running a web browser, without rewriting
them.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
