% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{start_fleet}
\alias{start_fleet}
\title{Starts the specified fleet}
\usage{
start_fleet(
  Name = "string"
)
}
\arguments{
\item{Name}{[required] The name of the fleet.}
}
\description{
Starts the specified fleet.
}
\examples{

}
