% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{stop_image_builder}
\alias{stop_image_builder}
\title{Stops the specified image builder}
\usage{
stop_image_builder(
  Name = "string"
)
}
\arguments{
\item{Name}{[required] The name of the image builder.}
}
\description{
Stops the specified image builder.
}
\examples{

}
