% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{tag_resource}
\alias{tag_resource}
\title{Adds or overwrites one or more tags for the specified AppStream 2}
\usage{
tag_resource(
  ResourceArn = "string",
  Tags = list(
    "string"
  )
)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{Tags}{[required] The tags to associate. A tag is a key-value pair (the value is optional). For example, \code{Environment=Test}, or, if you do not specify a value, \code{Environment=}.

If you do not specify a value, we set the value to an empty string.}
}
\description{
Adds or overwrites one or more tags for the specified AppStream 2.0 resource. You can tag AppStream 2.0 image builders, images, fleets, and stacks.
}
\details{
Each tag consists of a key and an optional value. If a resource already has a tag with the same key, this operation updates its value.

To list the current tags for your resources, use ListTagsForResource. To disassociate tags from your resources, use UntagResource.

For more information about tags, see \href{http://docs.aws.amazon.com/appstream2/latest/developerguide/tagging-basic.html}{Tagging Your Resources} in the \emph{Amazon AppStream 2.0 Developer Guide}.
}
\examples{

}
