% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{untag_resource}
\alias{untag_resource}
\title{Disassociates one or more specified tags from the specified AppStream 2}
\usage{
untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{TagKeys}{[required] The tag keys for the tags to disassociate.}
}
\description{
Disassociates one or more specified tags from the specified AppStream 2.0 resource.
}
\details{
To list the current tags for your resources, use ListTagsForResource.

For more information about tags, see \href{http://docs.aws.amazon.com/appstream2/latest/developerguide/tagging-basic.html}{Tagging Your Resources} in the \emph{Amazon AppStream 2.0 Developer Guide}.
}
\section{Accepted Parameters}{
\preformatted{untag_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

