% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{update_directory_config}
\alias{update_directory_config}
\title{Updates the specified Directory Config object in AppStream 2}
\usage{
update_directory_config(DirectoryName,
  OrganizationalUnitDistinguishedNames = NULL,
  ServiceAccountCredentials = NULL)
}
\arguments{
\item{DirectoryName}{The name of the Directory Config object.}

\item{OrganizationalUnitDistinguishedNames}{The distinguished names of the organizational units for computer accounts.}

\item{ServiceAccountCredentials}{The credentials for the service account used by the streaming instance to connect to the directory.}
}
\description{
Updates the specified Directory Config object in AppStream 2.0. This object includes the information required to join streaming instances to an Active Directory domain.
}
\examples{

}
