% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{associate_fleet}
\alias{associate_fleet}
\title{Associates the specified fleet with the specified stack}
\usage{
associate_fleet(FleetName, StackName)
}
\arguments{
\item{FleetName}{[required] The name of the fleet.}

\item{StackName}{[required] The name of the stack.}
}
\description{
Associates the specified fleet with the specified stack.
}
\section{Accepted Parameters}{
\preformatted{associate_fleet(
  FleetName = "string",
  StackName = "string"
)
}
}

