% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{delete_directory_config}
\alias{delete_directory_config}
\title{Deletes the specified Directory Config object from AppStream 2}
\usage{
delete_directory_config(DirectoryName)
}
\arguments{
\item{DirectoryName}{[required] The name of the directory configuration.}
}
\description{
Deletes the specified Directory Config object from AppStream 2.0. This object includes the information required to join streaming instances to an Active Directory domain.
}
\section{Accepted Parameters}{
\preformatted{delete_directory_config(
  DirectoryName = "string"
)
}
}

