% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{describe_images}
\alias{describe_images}
\title{Retrieves a list that describes one or more specified images, if the image names or image ARNs are provided}
\usage{
describe_images(
  Names = list(
    "string"
  ),
  Arns = list(
    "string"
  ),
  Type = "PUBLIC"|"PRIVATE"|"SHARED",
  NextToken = "string",
  MaxResults = 123
)
}
\arguments{
\item{Names}{The names of the public or private images to describe.}

\item{Arns}{The ARNs of the public, private, and shared images to describe.}

\item{Type}{The type of image (public, private, or shared) to describe.}

\item{NextToken}{The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.}

\item{MaxResults}{The maximum size of each page of results.}
}
\description{
Retrieves a list that describes one or more specified images, if the image names or image ARNs are provided. Otherwise, all images in the account are described.
}
\examples{

}
