% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_package.R
\docType{package}
\name{paws.appstream-package}
\alias{paws.appstream}
\alias{paws.appstream-package}
\title{paws.appstream: Amazon AppStream}
\description{
Amazon AppStream 2.0
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
