% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{stop_fleet}
\alias{stop_fleet}
\title{Stops the specified fleet}
\usage{
stop_fleet(Name)
}
\arguments{
\item{Name}{The name of the fleet.}
}
\description{
Stops the specified fleet.
}
\examples{

}
