% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{update_image_permissions}
\alias{update_image_permissions}
\title{Adds or updates permissions for the specified private image}
\usage{
update_image_permissions(Name, SharedAccountId, ImagePermissions)
}
\arguments{
\item{Name}{The name of the private image.}

\item{SharedAccountId}{The 12-digit ID of the AWS account for which you want add or update image permissions.}

\item{ImagePermissions}{The permissions for the image.}
}
\description{
Adds or updates permissions for the specified private image.
}
\examples{

}
