% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appstream_operations.R
\name{update_stack}
\alias{update_stack}
\title{Updates the specified fields for the specified stack}
\usage{
update_stack(Name, DisplayName = NULL, Description = NULL,
  StorageConnectors = NULL, DeleteStorageConnectors = NULL,
  RedirectURL = NULL, FeedbackURL = NULL, AttributesToDelete = NULL,
  UserSettings = NULL, ApplicationSettings = NULL)
}
\arguments{
\item{Name}{The name of the stack.}

\item{DisplayName}{The stack name for display.}

\item{Description}{The description for display.}

\item{StorageConnectors}{The storage connectors to enable.}

\item{DeleteStorageConnectors}{Deletes the storage connectors currently enabled for the stack.}

\item{RedirectURL}{The URL that users are redirected to after their streaming session ends.}

\item{FeedbackURL}{The URL that users are redirected to after they click the Send Feedback link. If no URL is specified, no Send Feedback link is displayed.}

\item{AttributesToDelete}{The stack attributes to delete.}

\item{UserSettings}{The actions that are enabled or disabled for users during their streaming sessions. By default, these actions are enabled.}

\item{ApplicationSettings}{The persistent application settings for users of a stack. When these settings are enabled, changes that users make to applications and Windows settings are automatically saved after each session and applied to the next session.}
}
\description{
Updates the specified fields for the specified stack.
}
\examples{

}
